# Update React 17→18, node 16&18→20, bump other dependency versions