--- a/.github/workflows/client-tests.yml
+++ b/.github/workflows/client-tests.yml
@@ -20,7 +20,7 @@ jobs:
         working-directory: client
     strategy:
       matrix:
-        node-version: [18.x]
+        node-version: [20.x]
         # See supported Node.js release schedule at https://nodejs.org/en/about/releases/
 
     steps: