--- a/.github/workflows/server-tests.yml
+++ b/.github/workflows/server-tests.yml
@@ -22,7 +22,7 @@ jobs:
         working-directory: server
     strategy:
       matrix:
-        node-version: [18.x]
+        node-version: [20.x]
         # See supported Node.js release schedule at https://nodejs.org/en/about/releases/
 
     steps: