--- a/client/src/index.jsx
+++ b/client/src/index.jsx
@@ -1,13 +1,14 @@
 import React from "react";
-import ReactDOM from "react-dom";
+import { createRoot } from 'react-dom/client';
 import { HashRouter as Router } from "react-router-dom";
 import App from "./App";
 
-ReactDOM.render(
+const container = document.getElementById('root');
+const root = createRoot(container);
+root.render(
   <React.StrictMode>
     <Router>
       <App />
     </Router>
-  </React.StrictMode>,
-  document.getElementById("root"),
+  </React.StrictMode>
 );