--- a/server/package.json
+++ b/server/package.json
@@ -29,7 +29,7 @@
     "@types/underscore": "^1.11.4",
     "debug-tree": "^0.1.3",
     "jest": "^27.4.7",
-    "nodemon": "^2.0.22",
+    "nodemon": "^3.1.7",
     "prettier": "^2.5.1",
     "supertest": "^6.2.2",
     "ts-jest": "^27.1.3",
@@ -37,11 +37,6 @@
     "typescript": "^4.5.4",
     "webpack": "^5.94.0",
     "webpack-cli": "^4.10.0"
-  },
-  "overrides": {
-    "body-parser": "^1.20.3",
-    "superagent": "^9.0.0",
-    "semver": "^7.5.2"
   },
   "scripts": {
     "docker:start": "docker-compose -p linkage-containers -f docker/docker-compose.yaml --env-file ./.env up -d && npm run docker:logs",