… client/src/data/templates.json