### Description

This addresses #405

### Testing

Pushing to staging first to test the app before merging to main.

### Impact on build time

On ARM64 the impact on build time of the server container is as follows.
(It seems that the SOO documentation is also created faster but I haven't monitored this precisely.)

Before: 
```
> docker build --no-cache -t lbl-api .
[+] Building 383.4s (36/36) FINISHED
...
 => [25/31] RUN node modelica-json/app.js -f modelica-buildings/Buildings/Templates -o json -d template-json                         149.4s
 => [26/31] WORKDIR /server                                                                                                            0.0s
 => [27/31] RUN npm i                                                                                                                  5.7s
 => [28/31] COPY . .                                                                                                                   0.1s
 => [29/31] RUN npm run parseTemplateJSON                                                                                             67.3s
```

After: 
```
> docker build --no-cache -t lbl-api .
[+] Building 294.0s (36/36) FINISHED
...
 => [25/31] RUN node modelica-json/app.js -f modelica-buildings/Buildings/Templates -o json -d template-json                         115.5s
 => [26/31] WORKDIR /server                                                                                                            0.0s
 => [27/31] RUN npm i                                                                                                                  3.0s
 => [28/31] COPY . .                                                                                                                   0.1s
 => [29/31] RUN npm run parseTemplateJSON                                                                                             41.7s
```